move file and push some tweaks
